+++ .storybook/config.js
@@ -1,7 +1,9 @@
-import { configure } from '@kadira/storybook';
+import { configure } from '@storybook/react';
+
+const req = require.context('../src', true, /\.stories\.jsx$/)
 
 function loadStories() {
-  require('../stories');
+  req.keys().forEach((filename) => req(filename))
 }
 
 configure(loadStories, module);
