+++ package.json
@@ -27,9 +27,9 @@
     "reactstrap": "^4.6.2"
   },
   "devDependencies": {
-    "@kadira/storybook": "^2.35.3",
     "@kadira/storybook-deployer": "^1.2.0",
     "@kadira/storyshots": "^2.1.0",
+    "@storybook/react": "3.1.3",
     "babel-cli": "^6.24.1",
     "babel-loader": "^7.0.0",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
